+++ extensions/easydict/src/releaseVersion/releaseNote.ts
@@ -2,7 +2,7 @@
  * @author: tisfeng
  * @createTime: 2022-07-03 22:10
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-10-26 11:02
+ * @lastEditTime: 2022-10-26 22:05
  * @fileName: releaseNote.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
@@ -16,6 +16,10 @@ export const releaseNote = `
 - 优化代码。
 - 更新 README。
 
+### 🐞 修复
+
+- 修复了最新 Raycast 1.42.0 版本中，arguments 为空会导致扩展崩溃的问题。
+
 #### 如果觉得这个扩展还不错，给个 [Star](https://github.com/tisfeng/Raycast-Easydict) ⭐️ 支持一下吧 (^-^)
 
 ---
@@ -24,4 +28,8 @@ export const releaseNote = `
 
 - Optimized code.
 - Updated README.
+
+### 🐞 Fixes
+
+- Fixed the problem that the latest Raycast 1.42.0 version will cause the extension to crash when arguments is empty.
 `;