+++ extensions/easydict/CHANGELOG.md
@@ -1,11 +1,13 @@
 # `Easydict` Changelog
 
-## [v2.6.1] - 2022-10-26
-
 ### 💎 Improvement
 
 - Optimized code.
 - Updated README.
+
+### 🐞 Fixes
+
+- Fixed the problem that the latest Raycast 1.42.0 version will cause the extension to crash when arguments is empty.
 
 ## [v2.6.0] - 2022-10-18